Refactor: give host_build_graph one signed counting domain
hbg mixed uint32_t and int32_t for the same quantities. task_local_id()
returned uint32_t while every consumer of a local id -- the task table,
completion_flags, the tensormap task chains, the fanin payload, the
allocator -- addresses slots with int32_t, so each boundary carried a
static_cast. The Definition header had the same split: its counts were
unsigned while the payload counts and task args they are compared against
were signed.

Everything that counts things is now int32_t: task_local_id() and the
make_*_task() encoders, GraphDefinition's task/edge/root/boundary/arg-pool
counts, the CSR offset arrays and the element indices into the argument
pools, GraphRecording and RecordedInGraphTask's counts and offsets, and
GRAPH_MAX_{TENSOR,SCALAR}_ARGS. Byte quantities stay unsigned: total_bytes,
execution_storage_bytes and the off_ fields are dimensioned against
UINT32_MAX by graph_layout_section and reach 4 GB by design, so narrowing
them would be a functional change. Every counting field is capped by
MAX_IN_GRAPH_TASKS times a per-task constant, the largest being edge_count
at 1024 x 1023, which clears INT32_MAX by three orders of magnitude.

sizeof is unchanged for every device-copied struct, so the wire layout does
not move.

Signing the pool offsets makes a negative value representable where the
unsigned span tests used to reject it as a huge one, and signed those tests
go blind: a negative offset is below tensor_arg_count, and subtracting it
widens the remaining span, so both `offset > count` and
`count > total - offset` pass. bind_graph_topology therefore rejects a
negative offset outright, and graph_layout_definition does the same for
every count and offset it reads -- which is also what makes its
remaining-span subtractions provably non-negative.
RejectsNegativeInGraphTaskArgOffset fails without that check.

Two other guards are tightened rather than fixed, since both were already
unreachable. graph_execution_localize gated task_count with `== 0`, which
reads as if it lets -1 through, but graph_execution_storage_layout behind it
has always rejected `task_count <= 0`; the gate now says `<= 0` itself and
bind_graph_topology re-checks the range it depends on.
graph_layout_definition bounded its running totals at UINT32_MAX while the
fields they land in are 32-bit and really cap at 32768 tensors / 16384
scalars, so neither the old bound nor the new INT32_MAX one can fire.

Stale comments describing the removed scheduler watermark are dropped from
the two schedulers and shared_memory.h.

+++ tests/ut/cpp/a5/test_hbg_submit_poison.cpp
@@ -139,7 +139,7 @@ TEST_F(HbgSubmitPoisonTest, EveryDeviceReadFieldIsWrittenOverPoison) {
         const ChipTaskSlotState &st = entry.slot;
 
         // Descriptor: the task id is written to this exact local id.
-        EXPECT_EQ(simpler::hbg::task_local_id(desc.task_id), static_cast<uint32_t>(local));
+        EXPECT_EQ(simpler::hbg::task_local_id(desc.task_id), local);
         // task_state is written at submit (reset_for_reuse skips it): PENDING for a
         // dispatchable task, COMPLETED for a pre-completed hidden-alloc. Either way a
         // real enum, never poison.
@@ -178,5 +178,5 @@ TEST_F(HbgSubmitPoisonTest, EveryDeviceReadFieldIsWrittenOverPoison) {
     // The consumer's fanin is written: two duplicate deps dedupe to one.
     const TaskPayload &cons_pl = tasks.storage_at(simpler::hbg::task_local_id(consumer.task_id())).payload;
     EXPECT_EQ(cons_pl.fanin_count, 1);
-    EXPECT_EQ(cons_pl.fanin_data()[0], static_cast<int32_t>(simpler::hbg::task_local_id(root.task_id())));
+    EXPECT_EQ(cons_pl.fanin_data()[0], simpler::hbg::task_local_id(root.task_id()));
 }
