+++ tests/ut/cpp/common/test_hbg_graph_recording_bounds.cpp
@@ -96,7 +96,7 @@ TEST_F(HbgGraphRecordingBoundsTest, RecordedTaskIsKeyedByItsIndexNotByTheRunsNum
         filler.add_input(boundary);
         ASSERT_TRUE(orch.submit_dummy_task(filler).task_id().is_valid()) << "filler task " << i;
     }
-    ASSERT_EQ(orch.task_allocator.active_count(), static_cast<int32_t>(MAX_IN_GRAPH_TASKS));
+    ASSERT_EQ(orch.task_allocator.active_count(), MAX_IN_GRAPH_TASKS);
 
     GraphTaskArgs boundary_args;
     boundary_args.add_input(boundary);
@@ -114,7 +114,7 @@ TEST_F(HbgGraphRecordingBoundsTest, RecordedTaskIsKeyedByItsIndexNotByTheRunsNum
     EXPECT_EQ(simpler::hbg::task_id_space(in_graph_task_id), simpler::hbg::TaskIdSpace::IN_GRAPH)
         << "a recorded task must not take a GLOBAL id: nothing resolves it against the task table, and its low "
            "field is what keys the recording's hazard map";
-    EXPECT_EQ(simpler::hbg::task_local_id(in_graph_task_id), 0u)
+    EXPECT_EQ(simpler::hbg::task_local_id(in_graph_task_id), 0)
         << "the first recorded task's low field is task index 0, independent of how many tasks the run has "
            "already allocated";
     EXPECT_LT(simpler::hbg::task_local_id(in_graph_task_id), MAX_IN_GRAPH_TASKS);
